Revert checkCallback function signature
Some idiot changed this signature in 367fb48224fe709bc19fa2fe43f88006e1927474.
Revert this because off_t is not the same as long long and the callback
signature has to be matched explicitly in C.

+++ checkisomd5.c
@@ -49,7 +49,7 @@ int user_bailing_out(void) {
     return 0;
 }
 
-static int outputCB(void *const co, const off_t offset, const off_t total) {
+static int outputCB(void *const co, const long long offset, const long long total) {
     struct progressCBData *const data = co;
     int gaugeval = -1;
