+++ libcheckisomd5.h
@@ -1,8 +1,6 @@
 #ifndef __LIBCHECKISOMD5_H__
 #define __LIBCHECKISOMD5_H__
 
-#include <sys/types.h>
-
 enum isomd5sum_status {
     ISOMD5SUM_FILE_NOT_FOUND = -2,
     ISOMD5SUM_CHECK_NOT_FOUND = -1,
@@ -12,7 +10,7 @@ enum isomd5sum_status {
 };
 
 /* For non-zero return value, check is aborted. */
-typedef int (*checkCallback)(void *, off_t offset, off_t total);
+typedef int (*checkCallback)(void *, long long offset, long long total);
 
 int mediaCheckFile(char *iso, checkCallback cb, void *cbdata);
 int mediaCheckFD(int isofd, checkCallback cb, void *cbdata);
